requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
click==8.1.2
click-plugins==1.1.1
cligj==0.7.2
Django==3.2
django-crispy-forms==1.14.0
django-leaflet==0.28.2
Fiona==1.8.21
geopandas==0.10.2
munch==2.5.0
numpy
pandas==1.3.5
psycopg2-binary==2.9.3
pyproj==3.3.1
python-dateutil==2.8.2
pytz==2022.1
Shapely==1.8.1.post1
six==1.16.0
sqlparse==0.4.2
